+++ src/contracts/jsonable.py
@@ -1,0 +1,9 @@
+import json
+from abc import ABC, abstractmethod
+
+
+class Jsonable(ABC):
+
+    @abstractmethod
+    def to_json(self):
+        pass