+++ src/tests/testable.py
@@ -1,16 +1,21 @@
 from abc import ABC, abstractmethod
 from src.utils.customlogger import CustomLogger, LoggerAdapter
 from datetime import datetime
+from src.contracts.jsonable import Jsonable
+from src.utils.data import BaseData
 import traceback
+import json
 
 
 class Testable(ABC):
     TAG: str = 'test'
     output_format: str = '{datetime} - {name} - {level} - {message}'
     _logger: LoggerAdapter
+    _save_result: bool
 
-    def __init__(self):
+    def __init__(self, save_result: bool = False):
         self._logger = CustomLogger().logger('test', self.TAG)
+        self._save_result = save_result
 
     def run(self):
         try:
@@ -18,6 +23,8 @@ class Testable(ABC):
             self._logger.debug('success')
             self._logger.debug('result:' + str(result))
             self.info(str(result))
+            if self._save_result:
+                self._save_json_file(result)
         except (Exception,):
             self._logger.error('fail')
             self._logger.error('ERROR:')
@@ -45,6 +52,13 @@ class Testable(ABC):
     def error(self, msg):
         self._logger.error(msg)
         self.__console('ERROR', msg)
+
+    def _save_json_file(self, result):
+        with open('test' + '_' + self.TAG + '.json', 'w+') as f:
+            if isinstance(result, str):
+                f.write(result)
+            else:
+                f.write(json.dumps(result, ensure_ascii=False, indent=4, sort_keys=True))
 
     @abstractmethod
     def handle(self):