+++ src/tests/opendart.py
@@ -10,7 +10,7 @@ class Opendart(Testable):
 
         service = OpenDartService()
         corp_code = service.get_corp_code_by_stock_code('005930')
-        single_data = service.get_single(corp_code.corp_code, '2021', service.Q1.value)
-        collect = AcntCollection(single_data['005930'])
+        single_data = service.get_single(corp_code.corp_code, '2021', service.Q1)
+        collect = single_data['005930']
 
-        return collect.to_json()
+        return collect