+++ src/utils/util.py
@@ -54,4 +54,4 @@ def config_ini(name: str) -> ConfigParser:
 
 
 def object_to_json(obj: object):
-    return json.dumps(obj, default=lambda o: o.__dict__, sort_keys=True, indent=4, ensure_ascii=False)
+    return json.dumps(obj, default=lambda o: o.__dict__, sort_keys=True, indent=2, ensure_ascii=False)