+++ CHANGELOG.md
@@ -16,6 +16,8 @@
     - src.tests.testable
     - test 관련 logger 정의
 - 기존 'modules' 패키지 'src'(으)로 변경
+- customlogger
+    - logs 및 logs/tests 폴더 없을 경우 생성 하게 수정
 
 **fix**